Enable usage of ssh keys present outside docker-machine
We would like to use keys present in ~/.ssh or in an ssh agent.
IOW let ssh send the known keys.
docker-machine enables this feature if the `Driver.SSHKeyPath` is empty
and if we don't use the native ssh client.

If the keypair name assigned doesn't exist in docker-machine, we assume
that this key is present outside docker-machine.